Remove `ModBlockFail`
Was only being used by `NoFail` now.
